Scheduler is now lock-free
Due to the previous heavy usage of Mutex inside the Scheduler there was
a starvation bug where a worker would lock the scheduler while sending
a file blocking the other workers on getting new jobs.

This has been solved adding a WorkerManager which handles all the
locking tasks, leaving the scheduler free of locks. FileStore had to be
patched in order to make it MT-safe, using internal mutability and OS
atomic operations for writing files.

The scheduler now only communicates via Channels an not via function
calls.

I took the opportunity to fix some other issues with the Status api and
the rendering of the scores in the curses ui.