Optimize cpu usage on network file transfers
Previously for sending a file to a remote channel it was serialized with
binenc. Profiling has revealed that this process was the main slowdown
during the evaluation. To avoid serializing raw file data a variant of
the file transfer has been implemented in an ad hoc way: the file
protocol messages has sent to the wire differently, especially the
Vec<u8> with the file content. First the size of the chunk is sent using
binenc codec, then the raw socket is used to write the buffer, skipping
the serialization part entirely.

The speedup is huge, from ~1min to ~28sec on the client side, using
same-host local tcp channel on a random task.

The CPU usage is dramatically reduced due to the skipped serialization.
Just the client went from 14.5s of userspace time to 0.28s.