Capturing stdout/stderr in ExecutionResult
With this patch it's possible to ask to the Executor to capture the
stdout and stderr of an Execution and send it directly with the
ExecutionResult structure.

This enormously simplifies the checker since all the needed resources
are in the same callback. It also simplifies the UIs since it's not
required to send new messages for the stdout/stderr of some executions.