Fix race condition when there is an early stop due to a failing callback
If one of the client's callbacks fail (e.g. because of a bugger checker)
the client exits without telling the executor that he left, leaving the
workers continuing their job.
This sometimes breaks the tests because the test finishes and the
tempdir gets dropped, removing the files the worker was using.

The fix is to exit more gracefully from the client, telling the executor
to stop and waiting for the sandbox threads.