Centralize server connection for clients and workers
This is finally adds support for default ports when connecting to a
remote host. The supported URL formats are:
- tcp://:password@host:port
- tcp://host:port
- tcp:://host
- host

Due to limitations in URL parsing the previous format (host:port) is
not supported anymore.
Domains are automatically resolved and if more addresses are found they
are tried in sequence until one responds. Handshake errors (wrong
password/wrong version/...) stop all the next tries.