Copying a file may cause a race condition if the program forks
If while copying the executable, task-maker forks (for example because
another sandbox start the execution), the open file descriptor is copied
to the other process and it won't be closed until the exec syscall
(std::fs::copy opens the files with FD_CLOEXEC). When the sandbox tries
to run the binary it may fail with ETXTBSY ("Text file busy") since it's
still open in the forked process.

The workaround is to avoid copying the file and hard-link it instead.
This may fail if the sandbox directory is in a different file system. In
this case try again copying and take the risk of ETXTBSY.