Fail if there are too many "managers"
For a IOI task the "managers" are: generator, validator, checker,
official solution, manager (of communication tasks).

For terry tasks the "managers" are: generator, validator, checker.

Previously some of the were just warnings, but the resulting behaviour
was way too random: the used manager was the first returned by glob, the
other was simply ignored.