Panic the sandbox runner if there is an unrecoverable error
There's little choice but to panic the group manager thread if there is
an error. This will still leave the system in a degraded state (where
the workers may look as busy, while they're stuck). At least now there
are some logs regarding the failure, previously the error was silently
ignored.