Support generators in communication tasks
Without this patch you couldn't have a generator in a communication task
because it was failing to find an "output generator". In fact the
solution cannot be used to produce the output files. This was simply
fixed by using the OutputGenerator::NotAvailable variant (which was
there for this reason, but unused).